refactor(editor): code mermaid nodeview 분리
